Finish relationship dynamics plot editor
