+++ src/main/python/plotlyst/view/widget/scene/relationship.py
@@ -21,110 +21,23 @@ from PyQt6.QtCore import pyqtSignal, Qt
 from PyQt6.QtGui import QResizeEvent, QIcon
 from PyQt6.QtWidgets import QWidget
 from overrides import overrides
-from qthandy import incr_font, incr_icon, vbox, vspacer, line, decr_font
+from qthandy import incr_font, incr_icon, vbox
 from qthandy.filter import VisibilityToggleEventFilter
-from qtmenu import MenuWidget
 
 from plotlyst.core.domain import StoryElement, CharacterAgency, Novel
 from plotlyst.env import app_env
 from plotlyst.service.cache import entities_registry
-from plotlyst.view.common import label, push_btn, tool_btn, columns, wrap, rows, ExclusiveOptionalButtonGroup, frame
+from plotlyst.view.common import label, push_btn, tool_btn
 from plotlyst.view.icons import IconRegistry, avatars
-from plotlyst.view.style.base import transparent_menu
-from plotlyst.view.widget.button import SelectorToggleButton
 from plotlyst.view.widget.display import MenuOverlayEventFilter
 from plotlyst.view.widget.input import RemovalButton
+from plotlyst.view.widget.relationship import AbstractRelationshipChangeDimensionPopup, DimensionSelectorButton, \
+    ModifierSelectorButton
 
 
-class _DimensionSelectorButton(SelectorToggleButton):
-    def __init__(self, dimension: str, icon: str, parent=None):
-        super().__init__(Qt.ToolButtonStyle.ToolButtonTextBesideIcon, minWidth=80, parent=parent)
-        self.dimension = dimension
-        self.iconName = icon
-        self.setText(dimension)
-        self.setIcon(IconRegistry.from_name(icon))
-
-
-class _ModifierSelectorButton(SelectorToggleButton):
-    def __init__(self, modifier: str, parent=None):
-        super().__init__(Qt.ToolButtonStyle.ToolButtonTextBesideIcon, minWidth=60, parent=parent)
-        self.modifier = modifier
-        self.setText(modifier)
-        decr_font(self)
-
-
-class RelationshipChangeDimensionPopup(MenuWidget):
+class RelationshipChangeDimensionPopup(AbstractRelationshipChangeDimensionPopup):
     def __init__(self, element: StoryElement, parent=None):
         super().__init__(parent)
-        transparent_menu(self)
-
-        self.wdgFrame = frame()
-        vbox(self.wdgFrame, 10)
-        self.wdgFrame.setProperty('white-bg', True)
-        self.wdgFrame.setProperty('large-rounded', True)
-
-        self.btnGroupDimensions = ExclusiveOptionalButtonGroup()
-        self.btnGroupModifiers = ExclusiveOptionalButtonGroup()
-
-        self.btnGroupDimensions.buttonClicked.connect(self._dimensionClicked)
-
-        self.wdgBond = rows(0)
-        self.wdgConflict = rows(0)
-        self.wdgCooperation = rows(0)
-
-        self.wdgEditor = columns(0, 8)
-        self.wdgEditor.layout().addWidget(self.wdgBond)
-        self.wdgEditor.layout().addWidget(self.wdgCooperation)
-        self.wdgEditor.layout().addWidget(self.wdgConflict)
-
-        self.wdgFrame.layout().addWidget(
-            label(
-                "Select in which dimension the relationship evolves—bond, cooperation, or conflict—or choose a more specific subtype within those categories",
-                description=True, wordWrap=True))
-        self.wdgFrame.layout().addWidget(self.wdgEditor)
-
-        btn = self.__initDimension('Bond', self.wdgBond, 'fa5s.hand-holding-heart')
-        incr_font(btn, 2)
-        self.wdgBond.layout().addWidget(line())
-        self.__initDimension('Love', self.wdgBond, 'fa5s.heart')
-        self.__initDimension('Friendship', self.wdgBond, 'ei.asl')
-        self.__initDimension('Family', self.wdgBond, 'ei.group-alt')
-
-        self.wdgBond.layout().addWidget(vspacer())
-
-        btn = self.__initDimension('Cooperation', self.wdgCooperation, 'fa5.handshake')
-        incr_font(btn, 2)
-        self.wdgCooperation.layout().addWidget(line())
-        self.__initDimension('Trust', self.wdgCooperation, 'fa5s.user-shield')
-        self.__initDimension('Alliance', self.wdgCooperation, 'fa5s.thumbs-up')
-        self.__initDimension('Respect', self.wdgCooperation, 'ri.award-fill')
-        self.__initDimension('Loyalty', self.wdgCooperation, 'ei.link')
-        self.wdgCooperation.layout().addWidget(vspacer())
-
-        btn = self.__initDimension('Conflict', self.wdgConflict, 'mdi.sword-cross')
-        incr_font(btn, 2)
-        self.wdgConflict.layout().addWidget(line())
-        self.__initDimension('Rivalry', self.wdgConflict, 'mdi6.trophy-outline')
-        self.__initDimension('Betrayal', self.wdgConflict, 'mdi6.knife')
-        self.__initDimension('Jealousy', self.wdgConflict, 'mdi.eye-circle-outline')
-        self.wdgConflict.layout().addWidget(vspacer())
-
-        self.wdgFrame.layout().addWidget(
-            wrap(label(
-                "Optionally select a modifier to reflect how the relationship dynamic has shifted",
-                description=True), margin_top=25), alignment=Qt.AlignmentFlag.AlignRight)
-
-        self.wdgModifiers = columns(0)
-        self.wdgFrame.layout().addWidget(self.wdgModifiers, alignment=Qt.AlignmentFlag.AlignRight)
-        self.__initModifier('Building')
-        self.__initModifier('Growing')
-        self.__initModifier('Strengthened')
-        self.__initModifier('Pressured')
-        self.__initModifier('Peaked')
-        self.__initModifier('Fading')
-        self.__initModifier('Broken')
-
-        self.addWidget(self.wdgFrame)
 
         if element.dimension:
             for btn in self.btnGroupDimensions.buttons():
@@ -139,29 +52,6 @@ class RelationshipChangeDimensionPopup(MenuWidget):
         else:
             for btn in self.btnGroupModifiers.buttons():
                 btn.setEnabled(False)
-
-    def _dimensionClicked(self):
-        checkedDim = self.btnGroupDimensions.checkedButton() is not None
-        if not checkedDim:
-            self.btnGroupModifiers.reset()
-
-        for btn in self.btnGroupModifiers.buttons():
-            btn.setEnabled(checkedDim)
-
-    def __initDimension(self, name: str, parent: QWidget, icon: str = '') -> _DimensionSelectorButton:
-        btn = _DimensionSelectorButton(name, icon)
-        self.btnGroupDimensions.addButton(btn)
-
-        parent.layout().addWidget(btn, alignment=Qt.AlignmentFlag.AlignCenter)
-
-        return btn
-
-    def __initModifier(self, modifier: str) -> _ModifierSelectorButton:
-        btn = _ModifierSelectorButton(modifier)
-        self.btnGroupModifiers.addButton(btn)
-        self.wdgModifiers.layout().addWidget(btn)
-
-        return btn
 
 
 class RelationshipChangeWidget(QWidget):
@@ -223,7 +113,7 @@ class RelationshipChangeWidget(QWidget):
         self._menu.installEventFilter(MenuOverlayEventFilter(self._menu))
         self._menu.exec()
 
-    def _dimensionChanged(self, btn: _DimensionSelectorButton):
+    def _dimensionChanged(self, btn: DimensionSelectorButton):
         if btn.isChecked():
             self._lblDimension.setText(btn.dimension)
             self._lblDimension.setIcon(IconRegistry.from_name(btn.iconName))
@@ -238,7 +128,7 @@ class RelationshipChangeWidget(QWidget):
             self.element.dimension = ''
             self.element.icon = ''
 
-    def _modifierChanged(self, btn: _ModifierSelectorButton):
+    def _modifierChanged(self, btn: ModifierSelectorButton):
         self._lblModifier.setVisible(btn.isChecked())
         if btn.isChecked():
             self._lblModifier.setText(f'[{btn.modifier}]')
